fix cypress types
