+++ cypress/support/slow-down.ts
@@ -1,3 +1,3 @@
 export const slowMs = 1000;
 export const fastMs = 100;
-export const defaultMs = process.env.CI ? slowMs : fastMs;
+export const defaultMs = process.env['CI'] ? slowMs : fastMs;
