+++ cypress/support/types/enums.ts
@@ -1,0 +1,14 @@
+export enum SkipOnboardingSlides {
+  No = 0,
+  Yes = 1
+}
+
+export enum HasBackedUp {
+  No = 0,
+  Yes = 1
+}
+
+export enum CheckIndexed {
+  No = 0,
+  Yes = 1
+}
